examples/plugins/parser/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = 'nomad-parser-plugin-example'
version = '1.0'
description = 'An example parser plugin for NOMAD.'
readme = "README.md"
authors = [{ name = "The NOMAD Authors" }]
license = { text = "Apache-2.0" }
dependencies = [
    "nomad-lab>=1.2.0"
]

[project.urls]
homepage = "https://github.com/nomad-coe/nomad-parser-example-plugin"

[project.optional-dependencies]
tests = [
    'pytest'
]

[tool.setuptools.packages.find]
include = ["nomadparserexample*"]

[tool.setuptools.package-data]
nomadschemaexample = ['*/nomad_plugin.yaml']
